Fixed some validations editions

+++ db/migrate/20110509175101_create_financial_assistance_options.rb
@@ -1,12 +1,12 @@
 class CreateFinancialAssistanceOptions < ActiveRecord::Migration
   def self.up
     create_table :financial_assistance_options do |t|
-      t.string :name
+      t.string  :name
       t.integer :financial_assistance_option_type_id
-      t.float :how_much
-      t.text :qualifications
-      t.string :source
-      t.string :website
+      t.integer :how_much
+      t.text    :qualifications
+      t.string  :source
+      t.string  :website
 
       t.timestamps
     end
